add a cli that can be triggered with --run-model <model> --prompt <prompt>
